requirements.txt
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dlib==19.24.4
et-xmlfile==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.51.0
install==1.3.5
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pathlib==1.0.1
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
